fix: fix api url and refactor frontend api files
